Optimize by merging delimit `RubyText` into `Ruby`
- `RubyDelimiterProcesser`内の`sutegana`メソッドを`Ruby`のDelimiterProcesserに編入
    - 合わせて、`use_rp_tag`の反映方法も変更
- `FuriganaOpenParser`クラスの65行目にあるDelimiterの定義のうち、`$canClose`をfalseに変更
- `RubyTextParser` からDelimterProcesserの処理を削除し、代わりに`RubyText`ノードを直接挿入
- `RubyText`ノードの継承元をAbstractInlineに変更
- `FuriganaExtension`の`register` から`RubyTextDelimterProcesser`の追加処理を削除
- 不要になったファイル、テストメソッドを削除
