+++ src/Delimiter/FuriganaDelimiterProcesser.php
@@ -41,14 +41,27 @@ class FuriganaDelimiterProcesser implements DelimiterProcessorInterface, Configu
         $this->config = $configuration;
     }
 
-    private function useRPTag(Ruby $ruby_node)
+    /**
+     * 捨て仮名を大文字に置換する関数。
+     * 引数にある置換フラグは実際にはプロパティで持たせる。
+     * 捨て仮名コードは以下の505行目から511行目までよりお借りしました。
+     *
+     * @see https://github.com/noisan/parsedown-rubytext/blob/master/lib/Parsedown/RubyTextTrait.php
+     *
+     * Copyright (c) 2015 Akihiro Yamanoi
+     * Released under the MIT license
+     * https://raw.githubusercontent.com/noisan/parsedown-rubytext/master/LICENSE
+     */
+    private function sutegana(string $ruby): string
     {
-        foreach ($ruby_node->iterator() as $node) {
-            if ($node instanceof RubyText) {
-                $node->insertBefore(new RubyParentheses('('));
-                $node->insertAfter(new RubyParentheses(')'));
-            }
-        }
+        $ruby_text_Sutegana = [
+            // 小書き文字をfromに、並字をtoに置く。ペアの要素順は合わせること
+            'from' => ['ぁ', 'ぃ', 'ぅ', 'ぇ', 'ぉ', 'っ', 'ゃ', 'ゅ', 'ょ', 'ゎ', 'ァ', 'ィ', 'ゥ', 'ェ', 'ォ', 'ヵ', 'ヶ', 'ッ', 'ャ', 'ュ', 'ョ', 'ヮ'],
+            'to' => ['あ', 'い', 'う', 'え', 'お', 'つ', 'や', 'ゆ', 'よ', 'わ', 'ア', 'イ', 'ウ', 'エ', 'オ', 'カ', 'ケ', 'ツ', 'ヤ', 'ユ', 'ヨ', 'ワ'],
+            // 小さいクなどは文字化けしてしまった
+        ];
+
+        return str_replace($ruby_text_Sutegana['from'], $ruby_text_Sutegana['to'], $ruby);
     }
 
     public function process(AbstractStringContainer $opener, AbstractStringContainer $closer, int $delimiterUse): void
@@ -56,14 +69,34 @@ class FuriganaDelimiterProcesser implements DelimiterProcessorInterface, Configu
         $node = new Ruby();
 
         $next = $opener->next();
-        while (null !== $next && $next !== $closer) {
+        while (null !== $next && !$next instanceof RubyText) {
             $tmp = $next->next();
             $node->appendChild($next);
             $next = $tmp;
         }
 
-        if ($this->config->get('furigana/use_rp_tag')) {
-            $this->useRPTag($node);
+        $rt = $next;
+        $next = $next->next();
+        $use_sutegana = $this->config->get('furigana/use_sutegana');
+        while (null !== $next && $next !== $closer) {
+            $tmp = $next->next();
+
+            if ($use_sutegana && $next instanceof AbstractStringContainer) {
+                $next->setLiteral($this->sutegana($next->getLiteral()));
+            }
+
+            $rt->appendChild($next);
+            $next = $tmp;
+        }
+
+        $node->appendChild($rt);
+
+        $use_rp_tag = $this->config->get('furigana/use_rp_tag');
+        foreach ($node->iterator() as $it) {
+            if ($use_rp_tag && $it instanceof RubyText) {
+                $it->insertBefore(new RubyParentheses('('));
+                $it->insertAfter(new RubyParentheses(')'));
+            }
         }
 
         $opener->insertAfter($node);
@@ -76,7 +109,7 @@ class FuriganaDelimiterProcesser implements DelimiterProcessorInterface, Configu
 
     public function getClosingCharacter(): string
     {
-        return '｜';
+        return '》';
     }
 
     public function getMinLength(): int
