+++ src/Node/RubyText.php
@@ -21,25 +21,10 @@ declare(strict_types=1);
 namespace JSW\Furigana\Node;
 
 use League\CommonMark\Node\Inline\AbstractInline;
-use League\CommonMark\Node\Inline\DelimitedInterface;
 
 /**
  * <rt>を担当するノード。
  */
-class RubyText extends AbstractInline implements DelimitedInterface
+final class RubyText extends AbstractInline
 {
-    public function __construct()
-    {
-        parent::__construct();
-    }
-
-    public function getOpeningDelimiter(): string
-    {
-        return '《';
-    }
-
-    public function getClosingDelimiter(): string
-    {
-        return '》';
-    }
 }
